+++ src/Calendrie.Future/Horology/UtcSystemClock.cs
@@ -44,7 +44,7 @@ public sealed class UtcSystemClock : IClock
     {
         var now = DateTime.UtcNow;
         // NB: the cast should always succeed.
-        int daysSinceZero = (int)TemporalArithmetic.DivideByTicksPerDay(now.Ticks);
+        int daysSinceZero = (int)TemporalArithmetic.DivideByTicksPerDay((ulong)now.Ticks);
         return new DayNumber(daysSinceZero);
     }
 }
